refactor(tools): inline time-window helpers into utils.go
