+++ internal/infra/mcp/tools/tool_k8s_list_top_http_errors_in_pods.go
@@ -145,14 +145,3 @@ func httpErrorsFilterString(clusterName, namespaceName, workloadType, workloadNa
 	}
 	return strings.Join(filters, ",")
 }
-
-// requestHasArg reports whether the caller explicitly supplied the named argument
-// (distinguishes "user passed empty string" from "user didn't pass the key at all").
-func requestHasArg(request mcp.CallToolRequest, name string) bool {
-	args, ok := request.Params.Arguments.(map[string]any)
-	if !ok {
-		return false
-	}
-	_, present := args[name]
-	return present
-}
